Fix security issue

+++ src/controllers/todo.controller.ts
@@ -4,12 +4,7 @@ import Todo from "../models/todo"
 
 const getTodos = async (req: Request, res: Response): Promise<void> => {
     try {
-        const {
-            params: { userId },
-            body,
-        } = req
-
-        const todos: ITodo[] = await Todo.find({ creator: userId }).sort({ status: 1, updatedAt: -1, createdAt: -1 })
+        const todos: ITodo[] = await Todo.find({ creator: req.userId }).sort({ status: 1, updatedAt: -1, createdAt: -1 })
         res.status(200).json({ todos })
     } catch (error) {
         throw error
@@ -28,11 +23,13 @@ const addTodo = async (req: Request, res: Response): Promise<void> => {
         })
 
         const newTodo: ITodo = await todo.save()
-        const allTodos: ITodo[] = await Todo.find({ creator: req.params.userId }).sort({ status: 1, updatedAt: -1, createdAt: -1 })
+        const allTodos: ITodo[] = await Todo.find({ creator: req.userId }).sort({ status: 1, updatedAt: -1, createdAt: -1 })
 
-        res
-            .status(201)
-            .json({ message: "Todo added", todo: newTodo, todos: allTodos })
+        res.status(201).json({ 
+            message: "Todo added", 
+            todo: newTodo, 
+            todos: allTodos 
+        })
     } catch (error) {
         throw error
     }
@@ -41,7 +38,7 @@ const addTodo = async (req: Request, res: Response): Promise<void> => {
 const updateTodo = async (req: Request, res: Response): Promise<void> => {
     try {
         const {
-            params: { todoId, userId },
+            params: { todoId },
             body,
         } = req
         const updateTodo: ITodo | null = await Todo.findByIdAndUpdate(
@@ -50,7 +47,7 @@ const updateTodo = async (req: Request, res: Response): Promise<void> => {
             },
             body
         )
-        const allTodos: ITodo[] = await Todo.find({ creator: userId }).sort({ status: 1, updatedAt: -1, createdAt: -1 })
+        const allTodos: ITodo[] = await Todo.find({ creator: req.userId }).sort({ status: 1, updatedAt: -1, createdAt: -1 })
         res.status(200).json({
             message: "Todo updated",
             todo: updateTodo,
@@ -63,10 +60,9 @@ const updateTodo = async (req: Request, res: Response): Promise<void> => {
 
 const deleteTodo = async (req: Request, res: Response): Promise<void> => {
     try {
-        const deletedTodo: ITodo | null = await Todo.findByIdAndRemove(
-            req.params.todoId
-        )
-        const allTodos: ITodo[] = await Todo.find({ creator: req.params.userId }).sort({ status: 1, updatedAt: -1, createdAt: -1 })
+        const deletedTodo: ITodo | null = await Todo.findByIdAndRemove(req.params.todoId)
+
+        const allTodos: ITodo[] = await Todo.find({ creator: req.userId }).sort({ status: 1, updatedAt: -1, createdAt: -1 })
         res.status(200).json({
             message: "Todo deleted",
             todo: deletedTodo,
@@ -79,10 +75,9 @@ const deleteTodo = async (req: Request, res: Response): Promise<void> => {
 
 const deleteAllTodos = async (req: Request, res: Response): Promise<void> => {
     try {
-        const deletedTodo = await Todo.deleteMany(
-            { creator: req.params.userId }
-        )
-        const allTodos: ITodo[] = await Todo.find({ creator: req.params.userId }).sort({ status: 1, updatedAt: -1, createdAt: -1 })
+        const deletedTodo = await Todo.deleteMany({ creator: req.userId })
+
+        const allTodos: ITodo[] = await Todo.find({ creator: req.userId }).sort({ status: 1, updatedAt: -1, createdAt: -1 })
         res.status(200).json({
             message: "Deleted Todos count: " + deletedTodo.deletedCount,
             todos: allTodos,
